+++ WhappenMyAzureVM.ps1
@@ -10,7 +10,8 @@
   Change Log
   ----------
   v1.00 Andy Ball 23\11\2016 Base Version 
-  v1.01 Andy Bakk 27\11\2016 If VM Name not found , output a list of all VMs
+  v1.01 Andy Ball 27\11\2016 If VM Name not found , output a list of all VMs
+  v1.02 Andy Ball 28\11\2016 Add CheckAllSubscription
 
   Backlog 
   --------
@@ -23,6 +24,9 @@
   
  .Parameter ResourceGroupName 
   Resource Group the VM resides in 
+
+  .Parameter CheckAllSubscriptions
+  If true , searches for VMName in all Subscriptions you have access to
   
  .Example
  WhappenMyAzureVM -VMName "MyVM" -ResourceGroupName "ItsResourceGroupName"
@@ -33,24 +37,77 @@ Function WhappenMyAzureVM
     Param
     (
      [Parameter (Mandatory = $true , Position = 0)] [string] $VMName,
-     [Parameter (Mandatory = $true , Position = 1)] [string] $ResourceGroupName
+     [Parameter (Mandatory = $false , Position = 1)] [string] $ResourceGroupName, 
+     [Parameter (Mandatory = $false , Position = 2)] [boolean] $CheckAllSubscriptions = $false
+
     )
 
-    # Yes i know about Find-AzureRMResource...
-    $VMs = Get-AzureRMVM 
-    $VM = $VMs | Where {$_.Name -eq $VMName -AND $_.ResourceGroupName -eq $ResourceGroupName}
+   #Requires -Modules AzureRM.Profile
+    
+    $AllVms = @()
+    $ErrorActionPreference = "Stop"
+    $VMFound = $false 
+    
+    # Get the current Subscription so we don't have to bother checking it again / can switch back to it if we search throough other subs
+    $CurrentRMContext = Get-AzureRMContext 
+    $OriginalSubscriptionName = $CurrentRMContext.Subscription.SubscriptionName
+    
+    Write-Host "Getting VMs in current Subscription = $OriginalSubscriptionName"
+    $VMs = Get-AzureRMVM | Select @{Name = "SubscriptionName" ; Expression = {$OriginalSubscriptionName}}, *
+    $VM = $VMs | Where {$_.Name -eq $VMName} # -AND $_.ResourceGroupName -eq $ResourceGroupName}
     If ($VM -eq $null)
     {
-        Write-Host "Cannot find VM = $VMName in resourceGroupName = $ResourceGroup, heres whats available"
-        $VMs | Select Name, ResourceGroupName | Sort Name, ResourceGroupName 
-       
-    }
-    Else
+        If ($CheckAllSubscriptions)
+            {
+                Write-Host "CheckAllSubscriptions is true, Getting List of Subscriptions"
+                # Add the current Subscriptions VMs to all 
+                $AllVMs += $VMs | Select @{Name = "SubscriptionName" ; Expression = {$OriginalSubscriptionName}}, *
+                $Subscriptions = Get-AzureRMSubscription | Where {$_.SubscriptionName -ne $OriginalSubscriptionName}
+          
+                ForEach ($Sub in $Subscriptions)
+                    {
+                            $VM = $null
+                            $SubscriptionName = $Sub.SubscriptionName 
+                            Write-Host "`tChecking SubscriptionName = $SubscriptionName"
+                            $ThisSub = Select-AzureRMSubscription -SubscriptionName $SubscriptionName
+                            $VMs = Get-AzureRMVM 
+                            $AllVMs += $VMs | Select @{Name = "SubscriptionName" ; Expression = {$SubscriptionName}}, *
+                            $VM = $VMs | Where {$_.Name -eq $VMName}
+                            If ($VM -ne $null)
+                                {
+                                    # Set to true so break out of the For
+                                    $VMFound = $True
+                                    break
+                                }
+      
+                    }
+
+                If ($SubscriptionName -ne $OriginalSubscriptionName)
+                    {
+                        Write-Host "Switching back from $SubscriptionName to $OriginalSubscriptionName"
+                        $ThisSub = Select-AzureRMSubscription -SubscriptionName $OriginalSubscriptionName
+                    }
+                # reuse this var so can use in output if not found
+                $Vms = $AllVMs
+            }
+
+    } #VM is null
+
+    # Finally load the web page if found .. 
+    If ($VMFound)
         {
+            Write-Host "`tVM Found !" -ForegroundColor Green 
             $URL = "https://portal.azure.com/#resource/" + $VM.Id + "/bootDiagnostics"
             Start-Process -FilePath $URL
         }
-
+    # Dump VM list if not.
+    Else
+        {
+            Write-Host "Cannot find VM = $VMName heres whats available :" 
+            $VMs | Select Name, ResourceGroupName, SubscriptionName | Sort Name, ResourceGroupName 
+        }
 }
 
-#WhappenMyAzureVM -VMName "myVMName" -ResourceGroupName "MyResourceGroupName"
+WhappenMyAzureVM -VMName "LBE-SV-IIST-001" -CheckAllSubscriptions $false 
+
+# -ResourceGroupName "MyResourceGroupName"